+++ build.gradle
@@ -206,7 +206,7 @@ ext {
 
 dependencies {
   implementation group: 'com.github.hmcts', name: 'et-data-model', version:'3.15.2'
-  implementation('com.github.hmcts:et-common:1.5.9') { transitive = false }
+  implementation('com.github.hmcts:et-common:860801c4a2') { transitive = false }
   implementation group: 'uk.gov.dwp.regex', name: 'postcode-validation', version: '1.1.0'
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web'
   implementation group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '3.1.9', {
