+++ src/main/java/uk/gov/hmcts/reform/et/syaapi/service/pdf/PdfUploadService.java
@@ -2,19 +2,13 @@ package uk.gov.hmcts.reform.et.syaapi.service.pdf;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.pdfbox.Loader;
-import org.apache.pdfbox.cos.COSName;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.pdmodel.PDDocumentCatalog;
-import org.apache.pdfbox.pdmodel.PDResources;
-import org.apache.pdfbox.pdmodel.font.PDType1Font;
-import org.apache.pdfbox.pdmodel.interactive.form.PDAcroForm;
-import org.apache.pdfbox.pdmodel.interactive.form.PDField;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
+import uk.gov.hmcts.ecm.common.exceptions.PdfServiceException;
 import uk.gov.hmcts.ecm.common.helpers.UtilHelper;
+import uk.gov.hmcts.ecm.common.service.pdf.PdfDecodedMultipartFile;
+import uk.gov.hmcts.ecm.common.service.pdf.PdfService;
 import uk.gov.hmcts.et.common.model.ccd.CaseData;
 import uk.gov.hmcts.et.common.model.ccd.items.GenericTseApplicationType;
 import uk.gov.hmcts.et.common.model.ccd.items.RespondentSumTypeItem;
@@ -30,36 +24,27 @@ import uk.gov.hmcts.reform.et.syaapi.service.utils.ClaimantTseUtil;
 import uk.gov.hmcts.reform.et.syaapi.service.utils.GenericServiceUtil;
 import uk.gov.hmcts.reform.idam.client.models.UserInfo;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static org.apache.commons.lang3.StringUtils.defaultIfEmpty;
 import static uk.gov.hmcts.et.common.model.ccd.types.citizenhub.ClaimantTse.APP_TYPE_MAP;
 import static uk.gov.hmcts.reform.et.syaapi.constants.EtSyaConstants.ENGLISH_LANGUAGE;
-import static uk.gov.hmcts.reform.et.syaapi.constants.EtSyaConstants.HELVETICA_PDFBOX_CHARACTER_CODE_1;
-import static uk.gov.hmcts.reform.et.syaapi.constants.EtSyaConstants.HELVETICA_PDFBOX_CHARACTER_CODE_2;
-import static uk.gov.hmcts.reform.et.syaapi.constants.EtSyaConstants.TIMES_NEW_ROMAN_PDFBOX_CHARACTER_CODE;
 import static uk.gov.hmcts.reform.et.syaapi.constants.EtSyaConstants.WELSH_LANGUAGE;
 
 /**
- * Uses {@link PdfMapperService} to convert a given case into a Pdf Document.
+ * Uses {@link PdfUploadService} to convert a given case into a Pdf Document.
  */
 @Slf4j
 @Service
 @RequiredArgsConstructor()
-@SuppressWarnings({"PMD.TooManyMethods", "PMD.ExcessiveImports"})
-public class PdfService {
+public class PdfUploadService {
 
-    private final PdfMapperService pdfMapperService;
+    private final PdfService pdfService;
     private final DocumentGenerationService documentGenerationService;
     @Value("${pdf.english}")
     public String englishPdfTemplateSource;
@@ -75,85 +60,6 @@ public class PdfService {
     private static final String PDF_FILE_TIKA_CONTENT_TYPE = "application/pdf";
     private static final String NOT_FOUND = "not found";
     private static final List<String> DOCUMENT_CHARS_TO_REPLACE = List.of("@", "/", "\\", "'");
-
-
-    /**
-     * Converts a {@link CaseData} class object into a pdf document
-     * using template (ver. ET1_1122)
-     *
-     * @param caseData  The data that is to be converted into pdf
-     * @param pdfSource The source location of the PDF file to be used as the template
-     * @return A byte array that contains the pdf document.
-     */
-    public byte[] convertCaseToPdf(CaseData caseData, String pdfSource) throws PdfServiceException {
-        byte[] pdfDocumentBytes;
-        try {
-            pdfDocumentBytes = createPdf(caseData, pdfSource);
-        } catch (IOException ioe) {
-            throw new PdfServiceException("Failed to convert to PDF", ioe);
-        }
-        return pdfDocumentBytes;
-    }
-
-    /**
-     * Populates a pdf document with data stored in the case data parameter.
-     *
-     * @param caseData  {@link CaseData} object with information in which to populate the pdf with
-     * @param pdfSource file name of the pdf template used to create the pdf
-     * @return a byte array of the generated pdf file.
-     * @throws IOException if there is an issue reading the pdf template
-     */
-    public byte[] createPdf(CaseData caseData, String pdfSource) throws IOException {
-        ClassLoader cl = Thread.currentThread().getContextClassLoader();
-        InputStream stream = ObjectUtils.isEmpty(cl) || StringUtils.isBlank(pdfSource) ? null
-            : cl.getResourceAsStream(pdfSource);
-        if (!ObjectUtils.isEmpty(stream)) {
-            try (PDDocument pdfDocument = Loader.loadPDF(
-                Objects.requireNonNull(stream))) {
-                PDResources resources = new PDResources();
-                resources.put(COSName.getPDFName(TIMES_NEW_ROMAN_PDFBOX_CHARACTER_CODE), PDType1Font.TIMES_ROMAN);
-                resources.put(COSName.getPDFName(HELVETICA_PDFBOX_CHARACTER_CODE_1), PDType1Font.HELVETICA);
-                resources.put(COSName.getPDFName(HELVETICA_PDFBOX_CHARACTER_CODE_2), PDType1Font.HELVETICA);
-                PDDocumentCatalog pdDocumentCatalog = pdfDocument.getDocumentCatalog();
-                PDAcroForm pdfForm = pdDocumentCatalog.getAcroForm();
-                pdfForm.setDefaultResources(resources);
-                for (Map.Entry<String, Optional<String>> entry : this.pdfMapperService.mapHeadersToPdf(caseData)
-                    .entrySet()) {
-                    String entryKey = entry.getKey();
-                    Optional<String> entryValue = entry.getValue();
-                    if (entryValue.isPresent()) {
-                        try {
-                            PDField pdfField = pdfForm.getField(entryKey);
-                            pdfField.setValue(entryValue.get());
-                        } catch (Exception e) {
-                            GenericServiceUtil.logException("Error while parsing PDF file for entry key \""
-                                                         + entryKey, caseData.getEthosCaseReference(), e.getMessage(),
-                                                            this.getClass().getName(), "createPdf");
-                        }
-                    }
-                }
-                ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-                pdfDocument.save(byteArrayOutputStream);
-                return byteArrayOutputStream.toByteArray();
-            } finally {
-                safeClose(stream, caseData);
-            }
-        }
-        safeClose(stream, caseData);
-        return new byte[0];
-    }
-
-    public static void safeClose(InputStream is, CaseData caseData) {
-        if (is != null) {
-            try {
-                is.close();
-            } catch (IOException e) {
-                GenericServiceUtil.logException("Input stream for the template PDF file was not closed: ",
-                                                caseData.getEthosCaseReference(), e.getMessage(),
-                                                "PDFServiceUtil", "safeClose");
-            }
-        }
-    }
 
     private static String createPdfDocumentNameFromCaseData(CaseData caseData,
                                                             String documentLanguage,
@@ -207,7 +113,7 @@ public class PdfService {
         userInfo) {
         List<PdfDecodedMultipartFile> files = new ArrayList<>();
         try {
-            byte[] pdfData = convertCaseToPdf(caseData, this.englishPdfTemplateSource);
+            byte[] pdfData = pdfService.convertCaseToPdf(caseData, this.englishPdfTemplateSource);
             if (ObjectUtils.isEmpty(pdfData)) {
                 throw new PdfServiceException(
                     "Failed to convert to PDF. English Template Not Found",
@@ -227,7 +133,7 @@ public class PdfService {
         }
         try {
             if (WELSH_LANGUAGE.equals(GenericServiceUtil.findClaimantLanguage(caseData))) {
-                byte[] pdfData = convertCaseToPdf(caseData, this.welshPdfTemplateSource);
+                byte[] pdfData = pdfService.convertCaseToPdf(caseData, this.welshPdfTemplateSource);
                 if (ObjectUtils.isEmpty(pdfData)) {
                     throw new PdfServiceException("Failed to convert to PDF. Welsh Template Not Found",
                                                   new NullPointerException());
