--- src/main/java/uk/gov/hmcts/reform/et/syaapi/service/utils/PdfMapperRespondentUtil.java
@@ -1,229 +1,0 @@
-package uk.gov.hmcts.reform.et.syaapi.service.utils;
-
-import org.apache.commons.lang3.StringUtils;
-import org.elasticsearch.common.Strings;
-import org.springframework.util.CollectionUtils;
-import org.springframework.util.ObjectUtils;
-import uk.gov.hmcts.et.common.model.ccd.Address;
-import uk.gov.hmcts.et.common.model.ccd.CaseData;
-import uk.gov.hmcts.et.common.model.ccd.items.RespondentSumTypeItem;
-import uk.gov.hmcts.et.common.model.ccd.types.RespondentSumType;
-import uk.gov.hmcts.reform.et.syaapi.models.AcasCertificatePdfFieldModel;
-import uk.gov.hmcts.reform.et.syaapi.models.RespondentPdfFieldModel;
-import uk.gov.hmcts.reform.et.syaapi.service.pdf.PdfServiceException;
-
-import java.util.Optional;
-import java.util.concurrent.ConcurrentMap;
-
-import static java.util.Optional.ofNullable;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.INVALID_NO_ACAS_REASON;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.MAX_NUMBER_OF_RESPONDENTS;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.NO;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.NO_LOWERCASE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.NUMERIC_FOUR_INT_VALUE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.NUMERIC_THREE_INT_VALUE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_CREATION_ERROR;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_MULTIPLE_RESPONDENTS_MIN_NUMBER;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_Q13_2_1_2_FORTH_RESPONDENT_ACAS_CERTIFICATE_CHECK_NO;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_Q2_3_1_2_FIRST_RESPONDENT_ACAS_CERTIFICATE_CHECK_NO;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_Q2_4_1_CLAIMANT_WORK_ADDRESS;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_Q2_4_2_CLAIMANT_WORK_POSTCODE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_Q2_5_MULTIPLE_RESPONDENTS;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_Q2_8_1_2_THIRD_RESPONDENT_ACAS_CERTIFICATE_CHECK_NO;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_ANOTHER_PERSON;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_EMPLOYER_ALREADY_IN_TOUCH;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_NO_POWER;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_UNFAIR_DISMISSAL;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.YES;
-
-public final class PdfMapperRespondentUtil {
-
-    private PdfMapperRespondentUtil() {
-        // Utility classes should not have a public or default constructor.
-    }
-
-    public static void putRespondents(CaseData caseData, ConcurrentMap<String, Optional<String>> printFields) {
-        try {
-            if (ObjectUtils.isEmpty(caseData) || CollectionUtils.isEmpty(caseData.getRespondentCollection())) {
-                putClaimantWorkAddress(caseData, printFields);
-            } else {
-                putFirstRespondent(caseData,
-                                   PdfTemplateRespondentFieldNamesEnum.FIRST_RESPONDENT.respondentPdfFieldModel,
-                                   printFields);
-                if (caseData.getRespondentCollection().size()
-                    >= PDF_TEMPLATE_MULTIPLE_RESPONDENTS_MIN_NUMBER) {
-                    printFields.put(
-                        PDF_TEMPLATE_Q2_5_MULTIPLE_RESPONDENTS,
-                        Optional.of(YES)
-                    );
-                    // Put Second Respondent
-                    putRespondent(
-                        caseData.getRespondentCollection().get(1),
-                        PdfTemplateRespondentFieldNamesEnum.SECOND_RESPONDENT.respondentPdfFieldModel,
-                        printFields
-                    );
-                }
-                // put third respondent if exists
-                if (caseData.getRespondentCollection().size() >= NUMERIC_THREE_INT_VALUE) {
-                    putRespondent(
-                        caseData.getRespondentCollection().get(2),
-                        PdfTemplateRespondentFieldNamesEnum.THIRD_RESPONDENT.respondentPdfFieldModel,
-                        printFields
-                    );
-                }
-                // put fourth respondent if exists
-                if (caseData.getRespondentCollection().size() >= NUMERIC_FOUR_INT_VALUE) {
-                    putRespondent(
-                        caseData.getRespondentCollection().get(3),
-                        PdfTemplateRespondentFieldNamesEnum.FORTH_RESPONDENT.respondentPdfFieldModel,
-                        printFields
-                    );
-                }
-                // put fifth respondent if exists
-                if (caseData.getRespondentCollection().size() == MAX_NUMBER_OF_RESPONDENTS) {
-                    putRespondent(
-                        caseData.getRespondentCollection().get(4),
-                        PdfTemplateRespondentFieldNamesEnum.FIFTH_RESPONDENT.respondentPdfFieldModel,
-                        printFields
-                    );
-                }
-            }
-        } catch (PdfServiceException pse) {
-            GenericServiceUtil.logException("Error while creating PDF file", caseData.getEthosCaseReference(),
-                                            pse.getMessage(), "PDFMapperRespondentUtil", "putRespondents");
-        }
-    }
-
-    private static void putFirstRespondent(CaseData caseData,
-                                          RespondentPdfFieldModel firstRespondentPdfFieldModel,
-                                          ConcurrentMap<String, Optional<String>> printFields)
-        throws PdfServiceException {
-        RespondentSumType firstRespondent = caseData.getRespondentCollection().get(0).getValue();
-        printFields.put(
-            firstRespondentPdfFieldModel.respondentNameFieldName(), ofNullable(firstRespondent.getRespondentName()));
-        putAddress(firstRespondent.getRespondentAddress(),
-                   firstRespondentPdfFieldModel.respondentAddressFieldName(),
-                   firstRespondentPdfFieldModel.respondentPostcodeFieldName(),
-                   printFields);
-        putAcasCertificateDetails(firstRespondent,
-                                  firstRespondentPdfFieldModel.respondentAcasCertificatePdfFieldModel(),
-                                  printFields);
-        putClaimantWorkAddress(caseData, printFields);
-    }
-
-    private static void putClaimantWorkAddress(CaseData caseData, ConcurrentMap<String, Optional<String>> printFields) {
-        if (!ObjectUtils.isEmpty(caseData)
-            && NO.equals(caseData.getClaimantWorkAddressQuestion())
-            && !ObjectUtils.isEmpty(caseData.getClaimantWorkAddress())) {
-            putAddress(caseData.getClaimantWorkAddress().getClaimantWorkAddress(),
-                       PDF_TEMPLATE_Q2_4_1_CLAIMANT_WORK_ADDRESS,
-                       PDF_TEMPLATE_Q2_4_2_CLAIMANT_WORK_POSTCODE,
-                       printFields);
-        }
-    }
-
-    private static void putRespondent(RespondentSumTypeItem respondentItem,
-                                      RespondentPdfFieldModel respondentPdfFieldModel,
-                                      ConcurrentMap<String, Optional<String>> printFields) throws PdfServiceException {
-        if (!ObjectUtils.isEmpty(respondentItem)) {
-            RespondentSumType respondent = respondentItem.getValue();
-            printFields.put(
-                respondentPdfFieldModel.respondentNameFieldName(),
-                ofNullable(respondent.getRespondentName())
-            );
-            putAddress(
-                respondent.getRespondentAddress(),
-                respondentPdfFieldModel.respondentAddressFieldName(),
-                respondentPdfFieldModel.respondentPostcodeFieldName(),
-                printFields
-            );
-            putAcasCertificateDetails(
-                respondent,
-                respondentPdfFieldModel.respondentAcasCertificatePdfFieldModel(),
-                printFields
-            );
-        }
-    }
-
-    private static void putAddress(Address address, String addressField, String postCodeField,
-                            ConcurrentMap<String, Optional<String>> printFields) {
-        if (!ObjectUtils.isEmpty(address)) {
-            printFields.put(
-                addressField,
-                ofNullable(PdfMapperServiceUtil.formatAddressForTextField(address))
-            );
-            printFields.put(
-                postCodeField,
-                ofNullable(PdfMapperServiceUtil.formatUkPostcode(address))
-            );
-        }
-    }
-
-    private static void putAcasCertificateDetails(RespondentSumType respondent,
-                                                  AcasCertificatePdfFieldModel acasCertificatePdfModel,
-                                                  ConcurrentMap<String, Optional<String>> printFields)
-        throws PdfServiceException {
-        if (StringUtils.isNotBlank(respondent.getRespondentAcasQuestion())
-            && YES.equals(respondent.getRespondentAcasQuestion())) {
-            printFields.put(acasCertificatePdfModel.getAcasCertificateCheckYesFieldName(),
-                            Optional.of(YES));
-            printFields.put(acasCertificatePdfModel.getAcasCertificateNumberFieldName(),
-                            ofNullable(respondent.getRespondentAcas())
-            );
-        } else {
-            putAcasCertificateNotFoundCheckbox(acasCertificatePdfModel, printFields);
-            if (!Strings.isNullOrEmpty(respondent.getRespondentAcasNo())) {
-                switch (respondent.getRespondentAcasNo()) {
-                    case PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_UNFAIR_DISMISSAL: {
-                        printFields.put(
-                            acasCertificatePdfModel.getNoAcasReasonUnfairDismissalFieldName(),
-                            Optional.of(YES)
-                        );
-                        break;
-                    }
-                    case PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_ANOTHER_PERSON: {
-                        printFields.put(
-                            acasCertificatePdfModel.getNoAcasReasonAnotherPersonFieldName(),
-                            Optional.of(YES)
-                        );
-                        break;
-                    }
-                    case PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_NO_POWER: {
-                        printFields.put(
-                            acasCertificatePdfModel.getNoAcasReasonNoPowerToConciliateFieldName(),
-                            Optional.of(YES)
-                        );
-                        break;
-                    }
-                    case PDF_TEMPLATE_REASON_NOT_HAVING_ACAS_EMPLOYER_ALREADY_IN_TOUCH: {
-                        printFields.put(
-                            acasCertificatePdfModel.getNoAcasReasonEmployerContactedFieldName(),
-                            Optional.of(YES)
-                        );
-                        break;
-                    }
-                    default: {
-                        throw new PdfServiceException(INVALID_NO_ACAS_REASON, new Exception(PDF_CREATION_ERROR));
-                    }
-                }
-            }
-        }
-    }
-
-    private static void putAcasCertificateNotFoundCheckbox(AcasCertificatePdfFieldModel acasCertificatePdfFieldModel,
-                                                           ConcurrentMap<String, Optional<String>> printFields) {
-        if (PDF_TEMPLATE_Q2_3_1_2_FIRST_RESPONDENT_ACAS_CERTIFICATE_CHECK_NO
-            .equals(acasCertificatePdfFieldModel.getAcasCertificateCheckNoFieldName())
-            || PDF_TEMPLATE_Q13_2_1_2_FORTH_RESPONDENT_ACAS_CERTIFICATE_CHECK_NO
-            .equals(acasCertificatePdfFieldModel.getAcasCertificateCheckNoFieldName())) {
-            printFields.put(acasCertificatePdfFieldModel.getAcasCertificateCheckNoFieldName(),
-                            Optional.of(NO_LOWERCASE));
-        } else if (PDF_TEMPLATE_Q2_8_1_2_THIRD_RESPONDENT_ACAS_CERTIFICATE_CHECK_NO
-            .equals(acasCertificatePdfFieldModel.getAcasCertificateCheckNoFieldName())) {
-            printFields.put(acasCertificatePdfFieldModel.getAcasCertificateCheckNoFieldName(), Optional.of(YES));
-        } else {
-            printFields.put(acasCertificatePdfFieldModel.getAcasCertificateCheckNoFieldName(), Optional.of(NO));
-        }
-    }
-
-}
