+++ src/test/java/uk/gov/hmcts/reform/et/syaapi/config/PostcodeToOfficeLookupTest.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import uk.gov.hmcts.ecm.common.configuration.PostcodeToOfficeMappings;
 
 import static org.assertj.core.api.Assertions.assertThat;
