+++ src/test/java/uk/gov/hmcts/reform/et/syaapi/helper/AssignCaseToLocalOfficeServiceTest.java
@@ -8,10 +8,10 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.dwp.regex.InvalidPostcodeException;
 import uk.gov.hmcts.ecm.common.model.helper.TribunalOffice;
+import uk.gov.hmcts.ecm.common.service.PostcodeToOfficeService;
 import uk.gov.hmcts.reform.et.syaapi.model.CaseTestData;
 import uk.gov.hmcts.reform.et.syaapi.models.CaseRequest;
 import uk.gov.hmcts.reform.et.syaapi.service.CaseOfficeService;
-import uk.gov.hmcts.reform.et.syaapi.service.PostcodeToOfficeService;
 
 import java.util.Optional;
