+++ src/test/java/uk/gov/hmcts/reform/et/syaapi/service/utils/TestConstants.java
@@ -1,10 +1,9 @@
 package uk.gov.hmcts.reform.et.syaapi.service.utils;
 
 import org.json.JSONObject;
-import uk.gov.hmcts.et.common.model.ccd.Address;
+import uk.gov.hmcts.ecm.common.service.pdf.PdfDecodedMultipartFile;
 import uk.gov.hmcts.et.common.model.ccd.types.DocumentType;
 import uk.gov.hmcts.et.common.model.ccd.types.UploadedDocumentType;
-import uk.gov.hmcts.reform.et.syaapi.service.pdf.PdfDecodedMultipartFile;
 import uk.gov.service.notify.SendEmailResponse;
 
 import java.util.ArrayList;
@@ -60,77 +59,18 @@ public final class TestConstants {
     public static final String FILE_NOT_EXISTS = "File does not exist!...";
     public static final String WELSH_LANGUAGE = "Welsh";
     public static final String ENGLISH_LANGUAGE = "English";
-    public static final String ADDRESS_LINE_1 = "AddressLine1";
-    public static final String ADDRESS_LINE_2 = "AddressLine2";
-    public static final String ADDRESS_LINE_3 = "AddressLine3";
-    public static final String POST_TOWN = "PostTown";
-    public static final String COUNTY = "County";
-    public static final String COUNTRY = "Country";
-    public static final String WORK_ADDRESS_LINE_1 = "WorkAddressLine1";
-    public static final String WORK_ADDRESS_LINE_2 = "WorkAddressLine2";
-    public static final String WORK_ADDRESS_LINE_3 = "WorkAddressLine3";
-    public static final String WORK_POST_TOWN = "WorkPostTown";
-    public static final String WORK_COUNTY = "WorkCounty";
-    public static final String WORK_COUNTRY = "WorkCountry";
-    public static final String WORK_POSTCODE = "SL6 2DE";
-    public static final String TEST_COMPANY_NAME = "Test Company Name";
-    public static final int INTEGER_NUMERIC_TWO = 2;
-    public static final int INTEGER_NUMERIC_THREE = 3;
-    public static final int INTEGER_NUMERIC_FOUR = 4;
-    public static final int INTEGER_NUMERIC_FIVE = 5;
-    public static final String STRING_NUMERIC_ONE = "1";
-    public static final String STRING_NUMERIC_TWO = "2";
-    public static final String STRING_NUMERIC_THREE = "3";
-    public static final String STRING_NUMERIC_FOUR = "4";
-    public static final String STRING_NUMERIC_FIVE = "5";
-    public static final String STRING_NUMERIC_SIX = "6";
-    public static final String POSTCODE = "SW1A 1AA";
-    public static final String REPRESENTATIVE_NAME = "Michael Mercury";
-    public static final String REPRESENTATIVE_REFERENCE = "Test Reference";
-    public static final String REPRESENTATIVE_ORGANISATION = "Test Organisation";
-    public static final String REPRESENTATIVE_PHONE_NUMBER = "03123456789";
-    public static final String REPRESENTATIVE_MOBILE_NUMBER = "07444518903";
-    public static final String REPRESENTATIVE_EMAIL = "test@gmail.com";
     public static final String EMPTY_STRING = "";
-    public static final String BLANK_STRING = "   ";
     public static final String NULL_STRING = null;
-    public static final Address NULL_ADDRESS = null;
-    public static final String TEST_ACAS = "R123456/78/90";
     public static final String CLAIMANT_IND_TYPE_NULL = "null";
     public static final String CLAIMANT_IND_TYPE_EMPTY = "empty";
     public static final String CLAIMANT_IND_TYPE_FILLED = "filled";
     public static final String INVALID_LANGUAGE = "Invalid Language";
     public static final boolean TRUE = true;
     public static final boolean FALSE = false;
-    public static final String CLAIM_DESCRIPTION = "Test Claim Description";
-    public static final String MR = "Mr";
-    public static final String MS = "Ms";
-    public static final String MRS = "Mrs";
-    public static final String MISS = "Miss";
-    public static final String ELIZABETH = "Elizabeth";
-    public static final String TAYLOR = "Taylor";
-    public static final String SWIFT = "Swift";
-    public static final String PREFER_NOT_TO_SAY = "Prefer not to say";
     public static final String YES = "Yes";
     public static final String NO = "No";
-    public static final String REASONABLE_ADJUSTMENT_DETAILS = "Wheelchair, Walking-Stick";
     public static final byte[] EMPTY_BYTE_ARRAY = new byte[0];
-    public static final String[] EMPTY_STRING_ARRAY = new String[0];
-    public static final String HEARING_ASSISTANCE = "Hearing Assistance";
-    public static final String VIDEO = "Video";
-    public static final String PHONE = "Phone";
     public static final String TEST_CASE_DATA_JSON_FILE = "requests/caseData.json";
-    public static final String WEEKS = "Weeks";
-    public static final String CLAIMANT_EMPLOYED_FROM = "2020-01-01";
-    public static final String CLAIMANT_EMPLOYED_TO = "2022-01-01";
-    public static final String CLAIMANT_OCCUPATION = "IT Manager";
-    public static final String AVERAGE_WEEKLY_HOURS = "37.5";
-    public static final String CLAIMANT_PAY_BEFORE_TAX = "45000";
-    public static final String CLAIMANT_PAY_AFTER_TAX = "40000";
-    public static final String CLAIMANT_PENSION_WEEKLY_CONTRIBUTION = "50";
-    public static final String CLAIMANT_BENEFITS_DETAIL = "Car";
-    public static final String NEW_PAY_BEFORE_TAX = "60000";
-    public static final String NO_LONGER_WORKING = "No longer working";
     public static final PdfDecodedMultipartFile PDF_DECODED_MULTIPART_FILE_NULL = new PdfDecodedMultipartFile(
         null,
         TestConstants.TEST_PDF_FILE_ORIGINAL_NAME,
@@ -145,19 +85,24 @@ public final class TestConstants {
         TestConstants.TEST_PDF_FILE_DOCUMENT_DESCRIPTION
     );
 
-    private static final String SEND_EMAIL_RESPONSE_BEGINNING = "{\n"
-        + "  \"id\": \"f30b2148-b1a6-4c0d-8a10-50109c96dc2c\",\n"
-        + "  \"reference\": \"TEST_EMAIL_ALERT\",\n"
-        + "  \"template\": {\n"
-        + "    \"id\": \"8835039a-3544-439b-a3da-882490d959eb\",\n"
-        + "    \"version\": \"3\",\n"
-        + "    \"uri\": \"TEST\"\n"
-        + "  },\n"
-        + "  \"content\": {\n";
-    private static final String SEND_EMAIL_RESPONSE_END = "    \"subject\": \"ET Test email created\",\n"
-        + "    \"from_email\": \"TEST@GMAIL.COM\"\n"
-        + "  }\n"
-        + "}\n";
+    private static final String SEND_EMAIL_RESPONSE_BEGINNING = """
+            {
+              "id": "f30b2148-b1a6-4c0d-8a10-50109c96dc2c",
+              "reference": "TEST_EMAIL_ALERT",
+              "template": {
+                "id": "8835039a-3544-439b-a3da-882490d959eb",
+                "version": "3",
+                "uri": "TEST"
+              },
+              "content": {
+            """;
+    public static final byte[] SAMPLE_BYTE_ARRAY = SEND_EMAIL_RESPONSE_BEGINNING.getBytes();
+    private static final String SEND_EMAIL_RESPONSE_END = """
+                "subject": "ET Test email created",
+                "from_email": "TEST@GMAIL.COM"
+              }
+            }
+            """;
     public static final SendEmailResponse INPUT_SEND_EMAIL_RESPONSE = new SendEmailResponse(
                         SEND_EMAIL_RESPONSE_BEGINNING
                         + "    \"body\": \"Dear test, Please see your detail as 123456789. Regards, ET Team.\",\n"
@@ -241,19 +186,7 @@ public final class TestConstants {
                                                + "4OS4gUmVnYXJkcywgRVQgVGVhbS4=\","
                                                + "\"confirm_email_before_download\":null,\"retention_period\""
                                                + ":null,\"is_csv\":false}");
-    public static final String CLAIMANT_PHONE_NUMBER_FIELD_NAME = "1.6 phone number";
-    public static final String CLAIMANT_MOBILE_NUMBER_FIELD_NAME = "1.7 mobile number";
-    public static final String CLAIMANT_EMAIL_FIELD_NAME = "1.9 email";
-    public static final String CLAIMANT_CONTACT_EMAIL_FIELD_NAME = "1.8 How should we contact you - Email";
-    public static final String OTHER = "Other";
-    public static final String OTHER_TITLE = "Other Title";
     public static final String MICHAEL = "Michael";
-    public static final String MERCURY = "Mercury";
-    public static final String MALE = "Male";
-    public static final String FEMALE = "Female";
-    public static final String EMAIL = "Email";
-    public static final String POST = "Post";
-    public static final String FAX = "Fax";
 
     private TestConstants() {
